requirements.txt
flask==3.1.3
python-dotenv==1.0.1
gunicorn==23.0.0
pytest==8.3.5
